auth-service/requirements.txt
fastapi
uvicorn[standard]
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
pydantic>=2,<3
pydantic-settings>=2,<3
asyncpg>=0.29,<1
sqlalchemy[asyncio]>=2,<3
redis[hiredis]>=5,<6